Migrate to transactional event store pattern for the User entity.
The change and the event are now committed in a single transaction.
This PR only does this change for the User entity, but we now store
a list of events along side the model. Any change the model appends
an event ot the list of events, and we commit them all at the end.