Remove null checks wherever possible and make the domain more pure
1. DTOs should be more explicit when they allow for null

As an example, in the FolderDTOs, setting to null would have previously
set the parentId to be the root folder. This is a hidden behavior that
is not well documented, so instead we introduced a proper discriminated
union to make the code more readable and also added an OpenAPI
description.

2. Remove usages of String.IsNullOrEmpty wherever possible

In the Infrastructure or Application layer, we will still occasionally
need this, but it shouldn't happen in the domain.

3. Remove usages of String.IsNullOrWhitespace wherever possible

In the Infrastructure or Application layer, we will still occasionally
need this, but it shouldn't happen in the domain.