Add input type validation when triggering runs
If the app has an input named 'Email' of type 'Email', we should
actually validate that the given input can be an Email.
Of course at run time they'll all be serialized into strings when
creating and calling the HTTP request.

This approach however does NOT work for multi-choice inputs. That
section needs to be reworked.