Move EntityFramework into the domain model
This is a tightly scoped side project, and as such will realistically
never need to migrate ORM frameworks. Decoupling EF from the domain
resulted in a lot of duplication - every HTTP request had *three* types
1) the DTO 2) the domain model 3) the Entity Framework model

Combining the two simplifies the codebase greatly and reduces unnecessary
boilerplate.

Of course, should we ever decide to move off of Entity Framework we will
rue the day we made this change, but that seems very unlikely given the
scope of this project.