Fix KeyValueList onblur bug
Autosave is tricky when you have two, related fields that both must
be filled. When moving from key to value, we would instantly get a
400 because the value was empty.

Instad, we need to have a wrapping div that combines both key and
value and only triggers the onBlur when neither the key nor the value
are being focused.